requirements-dev.txt
pytest>=8.0
tomlkit>=0.13
pyyaml>=6.0
